修复上传图片bug， 更换log

+++ app/src/main/java/com/xiaolian/amigo/data/network/model/dto/response/PersonalExtraInfoDTO.java
@@ -17,7 +17,7 @@ public class PersonalExtraInfoDTO {
     private List<BannerDTO> banners;
     private Integer bonusAmount;
     private Integer notifyAmount;
-    private Integer prepay;
+    private Double prepay;
     private Long repair;
     private Long lastRepairTime;
     private Notify urgentNotify;
